backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chroma>=0.2.0",
    "fastapi[all]>=0.116.1",
    "langchain-chroma>=0.2.5",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-openai>=0.3.30",
    "langchain-text-splitters>=0.3.9",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "sqlmodel>=0.0.24",
]
